+++ app/ElementsShell.ps1
@@ -1,0 +1,4 @@
+$modulePath = Join-Path -Path $PSScriptRoot -ChildPath '../src/AvePoint.Elements/AvePoint.Elements.psd1'
+Import-Module $modulePath -Force
+
+Start-AvptElementsShell